requirements.txt
tensorflow
transformers
streamlit
Pillow
torch
numpy
tf-keras